botifyx-bots
🍁support the creator by giving star to the repo which give him the 
      confidence for creating for such bot's for you guy's 

 🍁 CREATOR's ~ @Akuma_Rei_Kami
                       @ITSANIMEN
                       @Shadow_Slave_x

📢 Channel Description ~ https://t.me/BotifyX_Pro
(For your bot/channel where you release and host the auto-approval bot)
Auto Approval Bot is a Telegram utility designed to automatically approve user join requests for channels based on predefined conditions. It removes the need for manual approval, ensures fast entry for eligible users, and helps administrators manage large communities efficiently. Built for reliability, automation, and smooth access control.

💬 Support Group Description ~ https://t.me/BotifyX_support
(For discussion, help, and issues related to the bot)
Auto Approval Bot Support is the official support group for users of the Auto Approval Bot. This group is dedicated to help with setup, configuration, troubleshooting, updates, and feature discussions to ensure smooth and uninterrupted operation of the bot.